pyproject.toml
[project]
name = "pygame-visor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nitori", email = "nitori@chireiden.net" }
]
requires-python = ">=3.13"
dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pygame-ce>=2.5.3",
    "pytest>=8.3.5",
]
